1.2.6. updating filters to be extendable, and to work w urls correctly. adding index to taskmeta. fixing settings jump. Fixed (up) default assigned to and estimate.

+++ includes/class-db.php
@@ -1144,7 +1144,7 @@ abstract class Kanban_Db
 		}
 
 		// return the stored db version
-		return self::$installed_ver;
+		return (float) self::$installed_ver;
 	}
